Task 2/
The files in this directory download data from various online sources. All files need to be downloaded and the working directory in each file should be changed appropriately; as an example we use "../Github/Task 2/". All function files should be placed in the same directory, as indicated here.

The Golden Copy needed in Step 1can be downloaded directly from GLEIF: https://www.gleif.org/en/lei-data/gleif-golden-copy/download-the-golden-copy 

A "sample" output can be found in the appropriate folder.